test: to reduce the number of cfg(test) and simplify assembly, move tests from const_data to the tests directory
